requirements.txt
alembic==1.5.8
appdirs==1.4.4
astroid==2.5.3
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
flake8==3.9.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
regex==2021.4.4
six==1.15.0
SQLAlchemy==1.4.7
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
